Move workflows classes into separate files.
